Fix test collection errors and outdated model config assertions
Move load_builtin_tool from conftest.py to dedicated tool_loader.py to
resolve import collision with discord_gateway/tests/conftest.py, and
update model config test expectations for removed stockmark model and
changed claude-sonnet-4-5 context_length.
